+++ backend/src/main/java/com/example/demo/domain/post/dto/request/PostCreateRequest.java
@@ -3,8 +3,8 @@ package com.example.demo.domain.post.dto.request;
 import com.example.demo.domain.major.entity.Major;
 import com.example.demo.domain.post.entity.Post;
 import com.example.demo.domain.post.enums.PostStatus;
-import com.example.demo.domain.user.entity.Grade;
-import com.example.demo.domain.user.entity.Semester;
+import com.example.demo.domain.user.enums.Grade;
+import com.example.demo.domain.user.enums.Semester;
 import com.example.demo.domain.user.entity.User;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
