+++ backend/src/main/java/com/example/demo/domain/post/entity/Post.java
@@ -4,8 +4,8 @@ package com.example.demo.domain.post.entity;
 //import com.example.demo.domain.major.entity.Major;
 import com.example.demo.domain.major.entity.Major;
 import com.example.demo.domain.post.enums.PostStatus;
-import com.example.demo.domain.user.entity.Grade;
-import com.example.demo.domain.user.entity.Semester;
+import com.example.demo.domain.user.enums.Grade;
+import com.example.demo.domain.user.enums.Semester;
 import com.example.demo.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
